Moved moderation and gamification logic to scoutgame package
