--- apps/scoutgame/lib/moderation/checkIsBuilderBanned.ts
@@ -1,14 +1,0 @@
-import { prisma } from '@charmverse/core/prisma-client';
-
-export async function checkIsBuilderBanned(builderId: string) {
-  const builder = await prisma.scout.findUnique({
-    where: {
-      id: builderId,
-      bannedAt: {
-        not: null
-      }
-    }
-  });
-
-  return !!builder;
-}
