--- apps/scoutgame/lib/moderation/getStrikesForSeason.ts
@@ -1,13 +1,0 @@
-import { prisma } from '@charmverse/core/prisma-client';
-
-export async function getStrikesForSeason({ season }: { season: number }) {
-  const strikes = await prisma.builderStrike.count({
-    where: {
-      builderEvent: {
-        season
-      }
-    }
-  });
-
-  return strikes;
-}
