--- apps/scoutgame/lib/moderation/octokitClient.ts
@@ -1,5 +1,0 @@
-import { Octokit } from '@octokit/rest';
-
-export const octokit = new Octokit({
-  auth: process.env.SCOUTGAME_GITHUB_ACCOUNT_TOKEN
-});
