+++ packages/scoutgame/src/moderation/checkIsBuilderBanned.ts
@@ -1,0 +1,14 @@
+import { prisma } from '@charmverse/core/prisma-client';
+
+export async function checkIsBuilderBanned(builderId: string) {
+  const builder = await prisma.scout.findUnique({
+    where: {
+      id: builderId,
+      bannedAt: {
+        not: null
+      }
+    }
+  });
+
+  return !!builder;
+}
