+++ packages/scoutgame/src/moderation/getStrikesForSeason.ts
@@ -1,0 +1,13 @@
+import { prisma } from '@charmverse/core/prisma-client';
+
+export async function getStrikesForSeason({ season }: { season: number }) {
+  const strikes = await prisma.builderStrike.count({
+    where: {
+      builderEvent: {
+        season
+      }
+    }
+  });
+
+  return strikes;
+}
